% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rmd.R
\name{generate_about_rmd}
\alias{generate_about_rmd}
\title{Generate about.Rmd}
\usage{
generate_about_rmd(
  project_dir,
  name,
  author,
  email,
  affiliation,
  orcid,
  website,
  scheme,
  preset
)
}
\arguments{
\item{project_dir}{Character. Path to project root.}

\item{name}{Character. Project name.}

\item{author}{Character. Author name.}

\item{email}{Character or NULL. Author email.}

\item{affiliation}{Character or NULL. Affiliation.}

\item{orcid}{Character or NULL. ORCID iD.}

\item{website}{Character or NULL. Website URL.}

\item{scheme}{A scheme list.}

\item{preset}{A preset list.}
}
\value{
Invisibly returns the output file path.
}
\description{
Generate about.Rmd
}
\keyword{internal}
