% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_rmd.R
\name{generate_index_rmd}
\alias{generate_index_rmd}
\title{Generate index.Rmd}
\usage{
generate_index_rmd(project_dir, name, author, email, scheme, preset)
}
\arguments{
\item{project_dir}{Character. Path to project root.}

\item{name}{Character. Project name.}

\item{author}{Character. Author name.}

\item{email}{Character or NULL. Author email.}

\item{scheme}{A scheme list.}

\item{preset}{A preset list.}
}
\value{
Invisibly returns the output file path.
}
\description{
Generate index.Rmd
}
\keyword{internal}
