% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflowR-package.R
\docType{package}
\name{reflowR-package}
\alias{reflowR}
\alias{reflowR-package}
\title{reflowR: Themed Extensions for Workflowr Research Workflows}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Extends the workflowr package with themed color schemes, structured analysis pipeline templates, and utility scripts for reproducible research. Wraps workflowr::wflow_start() and adds five color schemes (clinical, basic research, code/statistics, special, and neutral), three depth presets for the Rmd pipeline (minimal, standard, extended), custom CSS/SCSS theming, SVG icons, ggplot2 theme functions, and a Quarto report template. All generated projects remain fully compatible with standard workflowr commands.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cttir/reflowR}
  \item \url{https://cttir.github.io/reflowR/}
  \item Report bugs at \url{https://github.com/cttir/reflowR/issues}
}

}
\author{
\strong{Maintainer}: R. Heller \email{raban.heller@uni-ulm.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})

Authors:
\itemize{
  \item R. Heller \email{raban.heller@uni-ulm.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID})
}

}
\keyword{internal}
